small tweaks
