+++ gc.go
@@ -1,5 +1,7 @@
 package gb
 
+import "path/filepath"
+
 // gc toolchain
 
 type gcToolchain struct {
@@ -9,4 +11,11 @@ type gcToolchain struct {
 
 type gcoption struct {
 	goos, goarch string
+}
+
+func (t *gcToolchain) Pack(afiles ...string) error {
+	args := []string{"r"}
+	args = append(args, afiles...)
+	dir := filepath.Dir(afiles[0])
+	return run(dir, t.pack, args...)
 }
